Trainer Views

+++ database/migrations/2022_06_06_220303_create_trainers_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('trainers', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->decimal('balance');
+            $table->integer('balance');
             $table->timestamps();
         });
     }
